// digitallog_and_notification_backend/src/upload/types/digital.ts
// Code generated by protoc-gen-ts_proto. DO NOT EDIT.
// versions:
//   protoc-gen-ts_proto  v2.12.0
//   protoc               v5.27.2
// source: digital.proto

/* eslint-disable */
import { GrpcMethod, GrpcStreamMethod } from "@nestjs/microservices";
import { Observable } from "rxjs";

export const protobufPackage = "digital";

export interface UploadRequest {
  table: string;
  fileName: string;
  contentType: string;
  chunk: Uint8Array;
}

export interface UploadResponse {
  statusCode: number;
  s3Url: string;
  error: string;
}

export interface DoesExistRequest {
  hash: string;
}

export interface DoesExistResponse {
  exists: boolean;
}

export const DIGITAL_PACKAGE_NAME = "digital";

export interface UploadServiceClient {
  upload(request: Observable<UploadRequest>): Observable<UploadResponse>;

  doesExist(request: DoesExistRequest): Observable<DoesExistResponse>;
}

export interface UploadServiceController {
  upload(request: Observable<UploadRequest>): Promise<UploadResponse> | Observable<UploadResponse> | UploadResponse;

  doesExist(request: DoesExistRequest): Promise<DoesExistResponse> | Observable<DoesExistResponse> | DoesExistResponse;
}

export function UploadServiceControllerMethods() {
  return function (constructor: Function) {
    const grpcMethods: string[] = ["doesExist"];
    for (const method of grpcMethods) {
      const descriptor: any = Reflect.getOwnPropertyDescriptor(constructor.prototype, method);
      GrpcMethod("UploadService", method)(constructor.prototype[method], method, descriptor);
    }
    const grpcStreamMethods: string[] = ["upload"];
    for (const method of grpcStreamMethods) {
      const descriptor: any = Reflect.getOwnPropertyDescriptor(constructor.prototype, method);
      GrpcStreamMethod("UploadService", method)(constructor.prototype[method], method, descriptor);
    }
  };
}

export const UPLOAD_SERVICE_NAME = "UploadService";